mlb_dl_streamlit/requirements.txt
python-mlb-statsapi==0.5.26
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
torch>=2.2.0
streamlit>=1.37.0
joblib>=1.3.0
